backend/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
python-dotenv==0.19.2
marshmallow==3.14.1
requests==2.26.0